Add 8-bit + bf16 quantization for ~3-4x faster inference on Apple Silicon
Optional `quantize` / `bf16` args on load_model() (and --quantize/--bf16 on
scripts/infer_indic_parler.py):

- quantize=8 — 8-bit-quantizes ONLY the AR decoder's nn.Linear layers (the ~89%
  inference bottleneck). Embeddings are intentionally skipped: decoder.py slices
  embed_positions.weight directly, so packing it breaks the forward pass with a
  broadcast error. The T5 encoder + DAC stay fp32.
- bf16 — halves activation + KV-cache memory traffic in the AR loop. Best paired
  with q8 (small weights + light activations).

Measured on an M4 Mac mini (indic-parler-tts), same sentence:
  CPU PyTorch 0.31x -> MLX fp32 0.49x -> q8 0.97x -> q8+bf16 1.28x realtime,
no audible quality loss. (8-bit beats 4-bit here on both speed and quality.)

Also casts bf16 audio -> fp32 before the numpy conversion in the return path.

Thanks @akashicMarga for the pure-MLX indic-parler implementation this builds on.

+++ models/indic_parler_tts/model.py
@@ -495,7 +495,7 @@ class IndicParlerTTS(nn.Module):
             f"AR={_t_ar:.2f}s ({last_step} steps, {_sps:.0f} step/s) | "
             f"DAC={_t_dac:.2f}s"
         )
-        return np.array(audio, dtype=np.float32)
+        return np.array(audio.astype(mx.float32), dtype=np.float32)
 
 
     def stream_generate(
@@ -607,7 +607,7 @@ class IndicParlerTTS(nn.Module):
                 # trim the context prefix — only yield the new portion
                 trim = len(_dac_ctx_frames) * 512
                 _dac_ctx_frames = new_frames[-_DAC_CTX:]
-                return np.array(audio, dtype=np.float32)[trim:]
+                return np.array(audio.astype(mx.float32), dtype=np.float32)[trim:]
 
             for step in range(1, max_steps):
                 last_step = step
